Splits out add todo reducer.
I'm not really sure if this is much of an improvement.
I'm structuring the folders of the application by feature,
yet the concept of todos is 'core' application data. I'm
not sure how much value there is in pulling the add todo
reducers into the TodoList container, in the future, there
may be multiple parts of the code base that add todos, so where
would it live then?

I guess it makes the tests smaller as you can test specifc
reducers..

Also, it may even make sense to take the todo fetching and pull
it into the TodoListGrid.